+++ src/main/java/home/example/board/domain/Comment.java
@@ -8,12 +8,12 @@ import java.time.LocalDateTime;
 @Data
 @Builder
 public class Comment {
-    private int comment_seq;
-    private Long post_seq;
+    private long comment_seq;
+    private long post_seq;
     private String content;
     private LocalDateTime insert_ts;
     private LocalDateTime delete_ts;
     private Integer delete_flag;
     private Integer parent_comment_seq;
-    private Long user_seq;
+    private long user_seq;
 }
